Fill + AutoFill -> FillStitch
